requirements.txt
pymongo==4.10.1
passlib==1.7.4
bcrypt==4.2.0
passlib==1.7.4
uvicorn==0.31.0
fastapi==0.115.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3